+++ packages/cms-base-nestjs-graphql-module/src/cms-base-graphql.module.ts
@@ -1,9 +1,5 @@
-import { Module, DynamicModule } from '@nestjs/common';
-import {
-  CMSBaseModule,
-  CMSBaseModuleOptionsDto,
-  CMSBaseModuleAsyncOptionsDto,
-} from '@rytass/cms-base-nestjs-module';
+import { Module, DynamicModule, Provider, Type } from '@nestjs/common';
+import { CMSBaseModule } from '@rytass/cms-base-nestjs-module';
 import { ArticleMutations } from './mutations/article.mutations';
 import { CategoryMutations } from './mutations/category.mutations';
 import { ArticleQueries } from './queries/article.queries';
@@ -13,10 +9,18 @@ import { ArticleDataLoader } from './data-loaders/article.dataloader';
 import { ArticleResolver } from './resolvers/article.resolver';
 import { BackstageArticleResolver } from './resolvers/backstage-article.resolver';
 import { ArticleSignatureResolver } from './resolvers/article-signature.resolver';
+import { BackstageCategoryResolver } from './resolvers/backstage-category.resolver';
+import { CMSGraphqlBaseModuleOptionsDto } from './typings/cms-graphql-base-root-module-options.dto';
+import { CMSGraphqlBaseModuleAsyncOptionsDto } from './typings/cms-graphql-base-root-module-async-options.dto';
+import { OptionProviders } from './constants/option-providers';
+import { CMS_BASE_GRAPHQL_MODULE_OPTIONS } from './typings/cms-graphql-base-providers';
+import { CMSGraphqlBaseModuleOptionFactory } from './typings/cms-graphql-base-root-module-option-factory';
 
 @Module({})
 export class CMSBaseGraphQLModule {
-  static forRootAsync(options: CMSBaseModuleAsyncOptionsDto): DynamicModule {
+  static forRootAsync(
+    options: CMSGraphqlBaseModuleAsyncOptionsDto,
+  ): DynamicModule {
     return {
       module: CMSBaseGraphQLModule,
       imports: [
@@ -25,6 +29,8 @@ export class CMSBaseGraphQLModule {
       ],
       exports: [CMSBaseModule],
       providers: [
+        ...this.createAsyncProvider(options),
+        ...OptionProviders,
         MemberDataLoader,
         ArticleDataLoader,
         ArticleResolver,
@@ -34,16 +40,22 @@ export class CMSBaseGraphQLModule {
         ArticleMutations,
         CategoryQueries,
         CategoryMutations,
+        BackstageCategoryResolver,
       ],
     };
   }
 
-  static forRoot(options?: CMSBaseModuleOptionsDto): DynamicModule {
+  static forRoot(options: CMSGraphqlBaseModuleOptionsDto): DynamicModule {
     return {
       module: CMSBaseGraphQLModule,
       imports: [CMSBaseModule.forRoot(options)],
       exports: [CMSBaseModule],
       providers: [
+        {
+          provide: CMS_BASE_GRAPHQL_MODULE_OPTIONS,
+          useValue: options,
+        },
+        ...OptionProviders,
         MemberDataLoader,
         ArticleDataLoader,
         ArticleResolver,
@@ -53,6 +65,49 @@ export class CMSBaseGraphQLModule {
         ArticleMutations,
         CategoryQueries,
         CategoryMutations,
+        BackstageCategoryResolver,
+      ],
+    };
+  }
+
+  private static createAsyncProvider(
+    options: CMSGraphqlBaseModuleAsyncOptionsDto,
+  ): Provider[] {
+    if (options.useExisting || options.useFactory) {
+      return [this.createAsyncOptionsProvider(options)];
+    }
+
+    return [
+      this.createAsyncOptionsProvider(options),
+      ...(options.useClass
+        ? [
+            {
+              provide: options.useClass,
+              useClass: options.useClass,
+            },
+          ]
+        : []),
+    ];
+  }
+
+  private static createAsyncOptionsProvider(
+    options: CMSGraphqlBaseModuleAsyncOptionsDto,
+  ): Provider {
+    if (options.useFactory) {
+      return {
+        provide: CMS_BASE_GRAPHQL_MODULE_OPTIONS,
+        useFactory: options.useFactory,
+        inject: options.inject || [],
+      };
+    }
+
+    return {
+      provide: CMS_BASE_GRAPHQL_MODULE_OPTIONS,
+      useFactory: async (optionsFactory: CMSGraphqlBaseModuleOptionFactory) =>
+        await optionsFactory.createCMSOptions(),
+      inject: [
+        (options.useExisting ||
+          options.useClass) as Type<CMSGraphqlBaseModuleOptionFactory>,
       ],
     };
   }