+++ packages/cms-base-nestjs-graphql-module/src/dto/category-multi-language-name.dto.ts
@@ -1,0 +1,10 @@
+import { Field, ObjectType } from '@nestjs/graphql';
+
+@ObjectType('CategoryMultiLanguageName')
+export class CategoryMultiLanguageNameDto {
+  @Field(() => String)
+  language: string;
+
+  @Field(() => String)
+  name: string;
+}