+++ packages/cms-base-nestjs-graphql-module/src/typings/cms-graphql-base-root-module-options.dto.ts
@@ -1,0 +1,16 @@
+import { CMSBaseModuleOptionsDto } from '@rytass/cms-base-nestjs-module';
+import { CustomFieldInput } from '../dto/custom-field.input';
+
+export interface CMSGraphqlBaseModuleOptionsDto
+  extends CMSBaseModuleOptionsDto {
+  mapArticleCustomFieldsToEntityColumns?: (
+    customFields: CustomFieldInput[],
+  ) =>
+    | Promise<Record<string, string | object>>
+    | Record<string, string | object>;
+  mapCategoryCustomFieldsToEntityColumns?: (
+    customFields: CustomFieldInput[],
+  ) =>
+    | Promise<Record<string, string | object>>
+    | Record<string, string | object>;
+}